# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — form-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}