# lib/api-zod/src/generated/types/videoJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — form-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoJobPlan } from './videoJobPlan';
import type { VideoJobStatus } from './videoJobStatus';
import type { VideoJobVideoType } from './videoJobVideoType';

export interface VideoJob {
  id: string;
  /** @nullable */
  title?: string | null;
  videoType: VideoJobVideoType;
  /** Duration in seconds (10, 30, 60, 180) */
  duration: number;
  plan: VideoJobPlan;
  status: VideoJobStatus;
  hasWatermark: boolean;
  /** @nullable */
  outputUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  imageCount?: number;
  clipCount?: number;
  createdAt: Date;
  updatedAt?: Date;
}