# lib/api-zod/src/generated/types/videoJobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — form-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoJobInputPlan } from './videoJobInputPlan';
import type { VideoJobInputVideoType } from './videoJobInputVideoType';

export interface VideoJobInput {
  title?: string;
  videoType: VideoJobInputVideoType;
  /** Duration in seconds */
  duration: number;
  plan: VideoJobInputPlan;
  images?: Blob[];
  clips?: Blob[];
}