# lib/api-zod/src/generated/types/videosSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VirJoy AI — form-based AI video creation API
 * OpenAPI spec version: 0.1.0
 */
import type { VideosSummaryByPlan } from './videosSummaryByPlan';
import type { VideosSummaryByStatus } from './videosSummaryByStatus';
import type { VideosSummaryByType } from './videosSummaryByType';

export interface VideosSummary {
  total: number;
  byStatus: VideosSummaryByStatus;
  byType: VideosSummaryByType;
  byPlan: VideosSummaryByPlan;
}